+++ packages/hashtag/src/hashtagStrategy.ts
@@ -1,5 +1,5 @@
 import { ContentBlock } from 'draft-js';
-import { extractHashtagsWithIndices } from './utils';
+import { extractHashtagsWithIndices } from './utils/extractHashtags';
 
 export default (
   contentBlock: ContentBlock,
@@ -8,7 +8,7 @@ export default (
   const text = contentBlock.getText();
   const results = extractHashtagsWithIndices(text);
 
-  results.forEach((hashtag) => {
+  results.forEach(hashtag => {
     callback(hashtag.indices[0], hashtag.indices[1]);
   });
 };